+++ src/utils/Enums.ts
@@ -171,10 +171,15 @@ export enum StringKey {
   edgeSwipeStartHint = constants.edgeSwipeStartHint,
   emailFieldHint = constants.emailFieldHint,
   image = constants.image,
+
   insufficientContentMessage = constants.insufficientContentMessage,
+  /** iOS only */
   insufficientContentTitle = constants.insufficientContentTitle,
+  /** @deprecated Use {@link insufficientContentMessage} instead. */
   invalidCommentMessage = constants.invalidCommentMessage,
+  /** @deprecated Use {@link insufficientContentTitle} instead. */
   invalidCommentTitle = constants.invalidCommentTitle,
+
   invalidEmailMessage = constants.invalidEmailMessage,
   invalidEmailTitle = constants.invalidEmailTitle,
   invocationHeader = constants.invocationHeader,